Add stream conformance tests for PipeStream
