+++ src/libraries/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.CreateClient.cs
@@ -10,7 +10,7 @@ namespace System.IO.Pipes.Tests
     /// <summary>
     /// Tests for the constructors for NamedPipeClientStream
     /// </summary>
-    public class NamedPipeTest_CreateClient : NamedPipeTestBase
+    public class NamedPipeTest_CreateClient
     {
         [Fact]
         public static void NullPipeName_Throws_ArgumentNullException()
@@ -152,6 +152,13 @@ namespace System.IO.Pipes.Tests
                         safeHandle.DangerousRelease();
                 }
             }
+        }
+
+        [Fact]
+        public void NamedPipeClientStream_InvalidHandleInerhitability()
+        {
+            AssertExtensions.Throws<ArgumentOutOfRangeException>("inheritability", () => new NamedPipeClientStream("a", "b", PipeDirection.Out, 0, TokenImpersonationLevel.Delegation, HandleInheritability.None - 1));
+            AssertExtensions.Throws<ArgumentOutOfRangeException>("inheritability", () => new NamedPipeClientStream("a", "b", PipeDirection.Out, 0, TokenImpersonationLevel.Delegation, HandleInheritability.Inheritable + 1));
         }
     }
 }
