+++ src/libraries/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.CurrentUserOnly.cs
@@ -10,26 +10,26 @@ namespace System.IO.Pipes.Tests
     /// <summary>
     /// Tests for the constructors for NamedPipeClientStream
     /// </summary>
-    public class NamedPipeTest_CurrentUserOnly : NamedPipeTestBase
+    public class NamedPipeTest_CurrentUserOnly
     {
         [Fact]
         public static void CreateClient_CurrentUserOnly()
         {
             // Should not throw.
-            new NamedPipeClientStream(".", GetUniquePipeName(), PipeDirection.InOut, PipeOptions.CurrentUserOnly).Dispose();
+            new NamedPipeClientStream(".", PipeStreamConformanceTests.GetUniquePipeName(), PipeDirection.InOut, PipeOptions.CurrentUserOnly).Dispose();
         }
 
         [Fact]
         public static void CreateServer_CurrentUserOnly()
         {
             // Should not throw.
-            new NamedPipeServerStream(GetUniquePipeName(), PipeDirection.InOut, 2, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly).Dispose();
+            new NamedPipeServerStream(PipeStreamConformanceTests.GetUniquePipeName(), PipeDirection.InOut, 2, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly).Dispose();
         }
 
         [Fact]
         public static void CreateServer_ConnectClient()
         {
-            string name = GetUniquePipeName();
+            string name = PipeStreamConformanceTests.GetUniquePipeName();
             using (var server = new NamedPipeServerStream(name, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
             {
                 using (var client = new NamedPipeClientStream(".", name, PipeDirection.InOut, PipeOptions.CurrentUserOnly))
@@ -43,7 +43,7 @@ namespace System.IO.Pipes.Tests
         [Fact]
         public static void CreateServer_ConnectClient_UsingUnixAbsolutePath()
         {
-            string name = Path.Combine("/tmp", GetUniquePipeName());
+            string name = Path.Combine("/tmp", PipeStreamConformanceTests.GetUniquePipeName());
             using (var server = new NamedPipeServerStream(name, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
             {
                 using (var client = new NamedPipeClientStream(".", name, PipeDirection.InOut, PipeOptions.CurrentUserOnly))
@@ -58,7 +58,7 @@ namespace System.IO.Pipes.Tests
         [InlineData(PipeOptions.CurrentUserOnly, PipeOptions.None)]
         public static void Connection_UnderSameUser_SingleSide_CurrentUserOnly_Works(PipeOptions serverPipeOptions, PipeOptions clientPipeOptions)
         {
-            string name = GetUniquePipeName();
+            string name = PipeStreamConformanceTests.GetUniquePipeName();
             using (var server = new NamedPipeServerStream(name, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, serverPipeOptions))
             using (var client = new NamedPipeClientStream(".", name, PipeDirection.InOut, clientPipeOptions))
             {
@@ -75,9 +75,9 @@ namespace System.IO.Pipes.Tests
         [Fact]
         public static void CreateMultipleServers_ConnectMultipleClients()
         {
-            string name1 = GetUniquePipeName();
-            string name2 = GetUniquePipeName();
-            string name3 = GetUniquePipeName();
+            string name1 = PipeStreamConformanceTests.GetUniquePipeName();
+            string name2 = PipeStreamConformanceTests.GetUniquePipeName();
+            string name3 = PipeStreamConformanceTests.GetUniquePipeName();
             using (var server1 = new NamedPipeServerStream(name1, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
             using (var server2 = new NamedPipeServerStream(name2, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
             using (var server3 = new NamedPipeServerStream(name3, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
@@ -101,7 +101,7 @@ namespace System.IO.Pipes.Tests
             {
                 tasks.Add(Task.Run(() =>
                 {
-                    var name = GetUniquePipeName();
+                    var name = PipeStreamConformanceTests.GetUniquePipeName();
                     using (var server = new NamedPipeServerStream(name, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.CurrentUserOnly))
                     {
                         using (var client = new NamedPipeClientStream(".", name, PipeDirection.InOut, PipeOptions.CurrentUserOnly))
@@ -126,7 +126,7 @@ namespace System.IO.Pipes.Tests
 
             try
             {
-                string pipeName = GetUniquePipeName();
+                string pipeName = PipeStreamConformanceTests.GetUniquePipeName();
                 for (var i = 0; i < pipeServers.Length; i++)
                 {
                     pipeServers[i] = new NamedPipeServerStream(
