--- src/libraries/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.Read.cs
@@ -1,89 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Runtime.InteropServices;
-using System.Threading.Tasks;
-using Xunit;
-
-namespace System.IO.Pipes.Tests
-{
-    public class NamedPipeTest_Read_ServerOut_ClientIn : PipeTest_Read
-    {
-        protected override ServerClientPair CreateServerClientPair()
-        {
-            ServerClientPair ret = new ServerClientPair();
-            string pipeName = GetUniquePipeName();
-            var writeablePipe = new NamedPipeServerStream(pipeName, PipeDirection.Out, 1, PipeTransmissionMode.Byte, PipeOptions.Asynchronous);
-            var readablePipe = new NamedPipeClientStream(".", pipeName, PipeDirection.In, PipeOptions.Asynchronous);
-
-            Task clientConnect = readablePipe.ConnectAsync();
-            writeablePipe.WaitForConnection();
-            clientConnect.Wait();
-
-            ret.readablePipe = readablePipe;
-            ret.writeablePipe = writeablePipe;
-            return ret;
-        }
-    }
-
-    public class NamedPipeTest_Read_ServerIn_ClientOut : PipeTest_Read
-    {
-        protected override ServerClientPair CreateServerClientPair()
-        {
-            ServerClientPair ret = new ServerClientPair();
-            string pipeName = GetUniquePipeName();
-            var readablePipe = new NamedPipeServerStream(pipeName, PipeDirection.In, 1, PipeTransmissionMode.Byte, PipeOptions.Asynchronous);
-            var writeablePipe = new NamedPipeClientStream(".", pipeName, PipeDirection.Out, PipeOptions.Asynchronous);
-
-            Task clientConnect = writeablePipe.ConnectAsync();
-            readablePipe.WaitForConnection();
-            clientConnect.Wait();
-
-            ret.readablePipe = readablePipe;
-            ret.writeablePipe = writeablePipe;
-            return ret;
-        }
-    }
-
-    public class NamedPipeTest_Read_ServerInOut_ClientInOut : PipeTest_Read
-    {
-        protected override ServerClientPair CreateServerClientPair()
-        {
-            ServerClientPair ret = new ServerClientPair();
-            string pipeName = GetUniquePipeName();
-            var readablePipe = new NamedPipeServerStream(pipeName, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.Asynchronous);
-            var writeablePipe = new NamedPipeClientStream(".", pipeName, PipeDirection.InOut, PipeOptions.Asynchronous);
-
-            Task clientConnect = writeablePipe.ConnectAsync();
-            readablePipe.WaitForConnection();
-            clientConnect.Wait();
-
-            ret.readablePipe = readablePipe;
-            ret.writeablePipe = writeablePipe;
-            return ret;
-        }
-
-        public override bool SupportsBidirectionalReadingWriting => true;
-    }
-
-    public class NamedPipeTest_Read_ServerInOut_ClientInOut_APMWaitForConnection : PipeTest_Read
-    {
-        protected override ServerClientPair CreateServerClientPair()
-        {
-            ServerClientPair ret = new ServerClientPair();
-            string pipeName = GetUniquePipeName();
-            var readablePipe = new NamedPipeServerStream(pipeName, PipeDirection.InOut, 1, PipeTransmissionMode.Byte, PipeOptions.Asynchronous);
-            var writeablePipe = new NamedPipeClientStream(".", pipeName, PipeDirection.InOut, PipeOptions.Asynchronous);
-
-            Task serverConnect = Task.Factory.FromAsync(readablePipe.BeginWaitForConnection, readablePipe.EndWaitForConnection, null);
-            writeablePipe.Connect();
-            serverConnect.Wait();
-
-            ret.readablePipe = readablePipe;
-            ret.writeablePipe = writeablePipe;
-            return ret;
-        }
-
-        public override bool SupportsBidirectionalReadingWriting => true;
-    }
-}
