--- src/libraries/System.IO.Pipes/tests/NamedPipeTests/NamedPipeTest.cs
@@ -1,18 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-using System.Security.Principal;
-using Xunit;
-
-namespace System.IO.Pipes.Tests
-{
-    public class NamedPipeTest_netstandard17 : NamedPipeTestBase
-    {
-        [Fact]
-        public void NamedPipeClientStream_InvalidHandleInerhitability()
-        {
-            AssertExtensions.Throws<ArgumentOutOfRangeException>("inheritability", () => new NamedPipeClientStream("a", "b", PipeDirection.Out, 0, TokenImpersonationLevel.Delegation, HandleInheritability.None - 1));
-            AssertExtensions.Throws<ArgumentOutOfRangeException>("inheritability", () => new NamedPipeClientStream("a", "b", PipeDirection.Out, 0, TokenImpersonationLevel.Delegation, HandleInheritability.Inheritable + 1));
-        }
-    }
-}
