+++ src/libraries/System.IO.Pipes/tests/PipeStream.ProtectedMethods.Tests.cs
